+++ appkit/src/main/java/org/ergoplatform/appkit/config/WalletConfig.java
@@ -1,4 +1,4 @@
-package org.ergoplatform.ergotool;
+package org.ergoplatform.appkit.config;
 
 /**
  * Parameters for working with wallet
